+++ R/flex_dashboard.R
@@ -6,33 +6,33 @@
 #'to its containing web page.
 #'
 #'@inheritParams rmarkdown::html_document
+#'@inheritParams rmarkdown::html_document_base
 #'
-#'@param fig_retina Scaling to perform for retina displays (defaults to 2).
-#'  Note that for flexdashboard enabling retina scaling provides for both
-#'  crisper graphics on retina screens but also much higher quality
-#'  auto-scaling of R graphics within flexdashboard containers.
+#'@param fig_retina Scaling to perform for retina displays (defaults to 2). Note
+#'  that for flexdashboard enabling retina scaling provides for both crisper
+#'  graphics on retina screens but also much higher quality auto-scaling of R
+#'  graphics within flexdashboard containers.
 #'
 #'@param fig_mobile Create an additional rendering of each R graphics figure
-#'  optimized for rendering on mobile devices oriented in portrait mode.
-#'  If \code{TRUE}, creates a figure which is 3.75 x 4.80 inches wide;
-#'  if \code{FALSE}, create no additional figure for mobile devices;
-#'  if a numeric vector of length 2, creates a mobile figure with the
-#'  specified width and height.
+#'  optimized for rendering on mobile devices oriented in portrait mode. If
+#'  \code{TRUE}, creates a figure which is 3.75 x 4.80 inches wide; if
+#'  \code{FALSE}, create no additional figure for mobile devices; if a numeric
+#'  vector of length 2, creates a mobile figure with the specified width and
+#'  height.
 #'
-#'@param favicon Path to graphic to be used as a favicon for the dashboard.
-#' Pass \code{NULL} to use no favicon.
+#'@param favicon Path to graphic to be used as a favicon for the dashboard. Pass
+#'  \code{NULL} to use no favicon.
 #'
 #'@param logo Path to graphic to be used as a logo for the dashboard. Pass
-#'  \code{NULL} to not include a logo. Note
-#'  that no scaling is performed on the logo image, so it should fit exactly
-#'  within the dimensions of the navigation bar (48 pixels high for the
-#'  default "cosmo" theme, other themes may have slightly different navigation
-#'  bar heights).
+#'  \code{NULL} to not include a logo. Note that no scaling is performed on the
+#'  logo image, so it should fit exactly within the dimensions of the navigation
+#'  bar (48 pixels high for the default "cosmo" theme, other themes may have
+#'  slightly different navigation bar heights).
 #'
 #'@param social Specify a character vector of social sharing services to
 #'  automatically add sharing links for them on the \code{navbar}. Valid values
-#'  are "twitter", "facebook", "google-plus", "linkedin", and "pinterest" (more
-#'  than one service can be specified).
+#'  are "twitter", "facebook", "linkedin", and "pinterest" (more than one
+#'  service can be specified).
 #'
 #'@param source_code URL for source code of dashboard (used primarily for
 #'  publishing flexdashboard examples). Automatically creates a \code{navbar}
@@ -40,8 +40,9 @@
 #'
 #'@param navbar Optional list of elements to be placed on the flexdashboard
 #'  navigation bar. Each element should be a list containing a \code{title}
-#'  and/or \code{icon} field, an \code{href} field. Optional fields \code{target}
-#'  (e.g. "_blank") and \code{align} ("left" or "right") are also supported.
+#'  and/or \code{icon} field, an \code{href} field. Optional fields
+#'  \code{target} (e.g. "_blank") and \code{align} ("left" or "right") are also
+#'  supported.
 #'
 #'@param orientation Determines whether level 2 headings are treated as
 #'  dashboard rows or dashboard columns.
@@ -51,16 +52,26 @@
 #'  natural height, scrolling the page if necessary.
 #'
 #'@param storyboard \code{TRUE} to use a storyboard layout scheme that places
-#'  each dashboard component in a navigable storyboard frame. When a
-#'  storyboard layout is used the \code{orientation} and \code{vertical_layout}
-#'  arguments are ignored. When creating a dashbaord with multiple pages you
-#'  should apply the `{.storyboard}` attribute to individual pages rather
-#'  than using the global \code{storyboard} option.
+#'  each dashboard component in a navigable storyboard frame. When a storyboard
+#'  layout is used the \code{orientation} and \code{vertical_layout} arguments
+#'  are ignored. When creating a dashbaord with multiple pages you should apply
+#'  the `{.storyboard}` attribute to individual pages rather than using the
+#'  global \code{storyboard} option.
 #'
-#'@param theme Visual theme ("default", "bootstrap", "cerulean", "journal",
-#'  "flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper",
-#'  "sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default"
-#'  is specified.
+#'@param theme One of the following:
+#'  * A [bslib::bs_theme()] object (or a list of [bslib::bs_theme()] argument values)
+#'  * Use this option to choose any [Bootstrap
+#'      version](https://rstudio.github.io/bslib/articles/bslib.html#bootstrap-versions),
+#'      [Bootswatch
+#'      theme](https://rstudio.github.io/bslib/articles/bslib.html#bootswatch-themes),
+#'      or implement a [custom
+#'      theme](https://rstudio.github.io/bslib/articles/bslib.html#custom-themes).
+#'    * In this case, any `.scss`/`.sass` files provided to the `css`
+#'      parameter may utilize the `theme`'s underlying Sass utilities
+#'      (e.g., variables, mixins, etc).
+#'  * A character string specifying a [Bootswatch 3](https://bootswatch.com/3/)
+#'    theme name (for backwards-compatibility). The "cosmo" theme is used when
+#'    "default" is specified.
 #'
 #'@param highlight Syntax highlighting style. Supported styles include
 #'  "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
@@ -69,14 +80,14 @@
 #'@param devel Enable development mode (used for development of the format
 #'  itself, not useful for users of the format).
 #'
-#'@param resize_reload Disable the auto-reloading behavior when the window is resized.
-#' Useful when debugging large flexdashboard applications and this functionality
-#' is not needed.
+#'@param resize_reload Disable the auto-reloading behavior when the window is
+#'  resized. Useful when debugging large flexdashboard applications and this
+#'  functionality is not needed.
 #'
-#'@param ... Unused
+#'@param ... Other arguments to [rmarkdown::html_document_base()].
 #'
 #'@details See the flexdashboard website for additional documentation:
-#'  \href{http://rmarkdown.rstudio.com/flexdashboard/}{http://rmarkdown.rstudio.com/flexdashboard/}
+#'  \href{https://pkgs.rstudio.com/flexdashboard/}{https://pkgs.rstudio.com/flexdashboard/}
 #'
 #' @examples
 #' \dontrun{
@@ -98,7 +109,6 @@ flex_dashboard <- function(fig_width = 6.0,
                            fig_retina = 2,
                            fig_mobile = TRUE,
                            dev = "png",
-                           smart = TRUE,
                            self_contained = TRUE,
                            favicon = NULL,
                            logo = NULL,
@@ -121,17 +131,9 @@ flex_dashboard <- function(fig_width = 6.0,
                            resize_reload = TRUE,
                            ...) {
 
-  # manage list of exit_actions (backing out changes to knitr options)
-  exit_actions <- list()
+  opts_orig <- NULL
   on_exit <- function() {
-    for (action in exit_actions)
-      try(action())
-  }
-
-  # function for resolving resources
-  resource <- function(name) {
-    system.file("rmarkdown/templates/flex_dashboard/resources", name,
-                package = "flexdashboard")
+    options(opts_orig)
   }
 
   # force self_contained to FALSE in devel mode
@@ -163,10 +165,9 @@ flex_dashboard <- function(fig_width = 6.0,
   fill_page <- identical(vertical_layout, "fill")
 
   # resolve theme
-  if (identical(theme, "default"))
-    theme <- "cosmo"
-  else if (identical(theme, "bootstrap"))
-    theme <- "default"
+  theme <- resolve_theme(theme)
+  # Set internal option so that gauge() can resolve accent colors at render-time
+  opts_orig <- c(opts_orig, options(flexdashboard.theme = theme))
 
   # resolve auto_reload
   if (resize_reload == 'no' | grepl("fa?l?s?e?", resize_reload, ignore.case = T))
@@ -187,17 +188,7 @@ flex_dashboard <- function(fig_width = 6.0,
 
   # force to fill it's container (unless the option is already set)
   if (is.na(getOption('DT.fillContainer', NA))) {
-    options(DT.fillContainer = TRUE)
-    exit_actions <- c(exit_actions, function() {
-      options(DT.fillContainer = NULL)
-    })
-  }
-
-  # request that DT auto-hide navigation (unless the option is already set)
-  if (is.na(getOption('DT.autoHideNavigation', NA))) {
-    exit_actions <- c(exit_actions, function() {
-      options(DT.autoHideNavigation = NULL)
-    })
+    opts_orig <- c(opts_orig, options(DT.fillContainer = TRUE))
   }
 
   # add hook to capture fig.width and fig.height and serialized
@@ -287,6 +278,14 @@ flex_dashboard <- function(fig_width = 6.0,
 
     args <- c()
 
+    # Restore the original options when the server stops
+    # running (instead of when render() is done executing)
+    if (is_shiny_runtime(runtime)) {
+      opts_orig2 <- opts_orig
+      shiny::onStop(function() { options(opts_orig2) })
+      opts_orig <<- NULL
+    }
+
     # initialize includes if needed
     if (is.null(includes))
       includes <- list()
@@ -303,60 +302,40 @@ flex_dashboard <- function(fig_width = 6.0,
     add_graphic("logo", logo)
     add_graphic("favicon", favicon)
 
-    # include flexdashboard.css and flexdashboard.js (but not in devel
-    # mode, in that case relative filesystem references to
-    # them are included in the template along with live reload)
+    # Include flexdashboard.js unless we're in devel mode.
+    # In that case, relative filesystem references to
+    # them are included in the template, along with live reload
     if (devel) {
       args <- c(args, pandoc_variable_arg("devel", "1"))
-      dashboardCss <- NULL
-      dashboardScript <- NULL
     } else {
-      if (fill_page) {
-        fillPageCss <- readLines(resource("fillpage.css"))
-      } else {
-        fillPageCss <- NULL
-      }
-
-      theme <- ifelse(identical(theme, "default"), "bootstrap", theme)
+      # It's important that this CSS is included this way (i.e., not a
+      # htmlDependency()) so that the storyboard container has a defined size
+      # when sly JS executes (#332).
       dashboardCss <- c(
         '<style type="text/css">',
-        readLines(resource("flexdashboard.css")),
-        readLines(resource(paste0("theme-", theme, ".css"))),
-        fillPageCss,
+        readLines(resource("storyboard.css")),
+        if (fill_page) readLines(resource("fillpage.css")),
         '</style>'
       )
+      dashboardCssFile <- tempfile(fileext = "html")
+      writeLines(dashboardCss, dashboardCssFile)
+      includes$in_header <- c(includes$in_header, dashboardCssFile)
 
-      dashboardScript <- c(
-        '<script type="text/javascript">',
-        readLines(resource("flexdashboard.js")),
-        '</script>'
-      )
+      dashboardScriptFile <- tempfile(fileext = ".html")
+      dashboardScript <- c('<script type="text/javascript">', readLines(resource("flexdashboard.js")), '</script>')
+      writeLines(dashboardScript, dashboardScriptFile)
+      includes$before_body <- c(includes$before_body, dashboardScriptFile)
     }
 
     # if there is no fig_mobile height and width then pass the default
     if (is.null(fig_mobile))
       fig_mobile <- default_fig_mobile
 
-    # css
-    if (!is.null(dashboardCss)) {
-      dashboardCssFile <- tempfile(fileext = "html")
-      writeLines(dashboardCss, dashboardCssFile)
-      args <- c(args, pandoc_include_args(in_header = dashboardCssFile))
-    }
-
-    # script
-    if (!is.null(dashboardScript)) {
-      dashboardScriptFile <- tempfile(fileext = ".html")
-      writeLines(dashboardScript, dashboardScriptFile)
-      includes$before_body <- c(includes$before_body, dashboardScriptFile)
-    }
-
     # dashboard init script
     dashboardInitScript <- c(
        '<script type="text/javascript">',
        '$(document).ready(function () {',
        '  FlexDashboard.init({',
-       paste0('    theme: "', theme, '",'),
        paste0('    fillPage: ', ifelse(fill_page,'true','false'), ','),
        paste0('    orientation: "', orientation, '",'),
        paste0('    storyboard: ', ifelse(storyboard,'true','false'), ','),
@@ -365,6 +344,25 @@ flex_dashboard <- function(fig_width = 6.0,
        paste0('    defaultFigWidthMobile: ', figSizePixels(fig_mobile[[1]]), ','),
        paste0('    defaultFigHeightMobile: ', figSizePixels(fig_mobile[[2]]), ','),
        paste0('    resize_reload: ', ifelse(resize_reload,'true','false')),
+       '  });',
+       '  var navbar = $(".navbar").first();',
+       '  var body = $("body").first();',
+       '  var sidebar = $(".section.sidebar").first();',
+       '  function addNavbarPadding() {',
+       '    var navHeight = navbar.outerHeight();',
+       '    body.css("padding-top", (navHeight + 8) + "px");',
+       '    sidebar.css("top", navHeight + "px");',
+       # Trigger browser resize in a way that's supported on IE11
+       # https://developer.mozilla.org/en-US/docs/Web/Events/Creating_and_triggering_events#the_old-fashioned_way
+       # This is primarily here to make sure htmlwidgets (or anything else with
+       # resize callback) can size itself appropriately after padding is added
+       '    var resizeEvent = window.document.createEvent("UIEvents");',
+       '    resizeEvent.initUIEvent("resize", true, false, window, 0);',
+       '    window.dispatchEvent(resizeEvent);',
+       '  }',
+       '  if (!window.Shiny) setTimeout(addNavbarPadding, 100);',
+       '  $(document).on("shiny:idle", function() {',
+       '    setTimeout(addNavbarPadding, 50);',
        '  });',
        '});',
        '</script>'
@@ -395,9 +393,17 @@ flex_dashboard <- function(fig_width = 6.0,
                                         before_body = includes$before_body,
                                         after_body = includes$after_body))
 
-    # additional user css
-    for (css_file in css)
-      args <- c(args, "--css", pandoc_path_arg(css_file))
+    # html_document_base gained a css argument in v2.7.7
+    # (which also handles scss/sass files), so only do the
+    # CSS -> Pandoc conversion if these are css files
+    if (!is_installed("rmarkdown", "2.7.7")) {
+      for (css_file in css) {
+        if (grepl("\\.s[ac]ss$", css_file)) {
+          stop("Compilation of Sass -> CSS requires rmarkdown version 2.7.7 or higher")
+        }
+        args <- c(args, "--css", pandoc_path_arg(css_file))
+      }
+    }
 
     args
   }
@@ -415,12 +421,39 @@ flex_dashboard <- function(fig_width = 6.0,
   extra_dependencies <- append(extra_dependencies,
                                navbar_dependencies(navbar))
 
+  if (is.character(theme)) {
+    extra_dependencies <- append(extra_dependencies, list(valueBoxStaticAccentCSS(theme)))
+  }
+
   # depend on featherlight and prism for embedded source code
   if (source_code_embed(source_code)) {
     extra_dependencies <- append(extra_dependencies,
                                  list(html_dependency_jquery(),
                                       html_dependency_featherlight(),
                                       html_dependency_prism()))
+  }
+
+  if (is_bs_theme(theme)) {
+    if (!is_installed("rmarkdown", "2.7.1")) {
+      stop("Using a {bslib} theme requires rmarkdown v2.7.1 or higher")
+    }
+
+    # Attach the dynamic CSS dependency to the theme so that the dependency
+    # is restyled if and when `session$setCurrentTheme()` gets called
+    flexdb_css <- bslib::bs_dependency_defer(html_dependencies_flexdb)
+    theme <- bslib::bs_bundle(theme, sass::sass_layer(html_deps = flexdb_css))
+
+    # If $navbar-bg wasn't specified by user, default it to $primary
+    # (instead of $dark, since the template has .navbar-inverse)
+    navbar_bg <- grepl("$navbar-bg:", sass::as_sass(theme), fixed = TRUE)
+    if (!navbar_bg) {
+      theme <- bslib::bs_add_variables(
+        theme, primary = unname(getSassAccentColors(theme, "primary")),
+        "navbar-bg" = "$primary"
+      )
+    }
+  } else {
+    extra_dependencies <- append(extra_dependencies, html_dependencies_flexdb(theme))
   }
 
   # return format
@@ -435,13 +468,14 @@ flex_dashboard <- function(fig_width = 6.0,
     pre_knit = pre_knit,
     pre_processor = pre_processor,
     on_exit = on_exit,
-    base_format = html_document_base(smart = smart, theme = theme,
+    base_format = html_document_base(theme = theme,
                                      self_contained = self_contained,
                                      lib_dir = lib_dir, mathjax = mathjax,
                                      template = "default",
                                      pandoc_args = pandoc_args,
                                      bootstrap_compatible = TRUE,
                                      extra_dependencies = extra_dependencies,
+                                     css = css,
                                      ...)
   )
 }
@@ -475,6 +509,9 @@ source_code_embed_args <- function(source_file) {
 
   # embed it
   if (length(code) > 0) {
+
+    # escape in case there is html in code
+    code <- htmlEscape(code)
 
     # ensure we don't start with an emtpy line
     code[[1]] <- paste0(
@@ -534,7 +571,6 @@ navbar_links <- function(social, source_code) {
       menu$items <- list(
         list(title = "Twitter", icon = "fa-twitter"),
         list(title = "Facebook", icon = "fa-facebook"),
-        list(title = "Google+", icon = "fa-google-plus"),
         list(title = "LinkedIn", icon = "fa-linkedin"),
         list(title = "Pinterest", icon = "fa-pinterest")
       )
@@ -617,3 +653,48 @@ storyboard_dependencies <- function(source = NULL) {
 }
 
 
+html_dependencies_flexdb <- function(theme) {
+  name <- "flexdashboard-css"
+  version <- get_package_version("flexdashboard")
+
+  if (is.character(theme)) {
+    if (identical(theme, "default")) {
+      theme <- "bootstrap"
+    }
+    dep <- htmlDependency(
+      name = name, version = version,
+      src = "www/flex_dashboard",
+      package = "flexdashboard",
+      stylesheet = c(
+        "flexdashboard.min.css",
+        paste0("theme-", theme, ".css")
+      )
+    )
+    return(list(dep))
+  }
+
+  if (bslib::is_bs_theme(theme)) {
+    dep <- bslib::bs_dependency(
+      sass::sass_file(resource("flexdashboard.scss")),
+      theme = theme, name = name, version = version,
+      cache_key_extra = version
+    )
+    return(list(dep))
+  }
+
+  if (!is.null(theme)) {
+    warning("Didn't recognize a theme object with class: ", class(theme))
+  }
+
+  NULL
+}
+
+# function for resolving resources
+resource <- function(name) {
+  system_file("www/flex_dashboard", name, package = "flexdashboard")
+}
+
+# copied from rmarkdown:::is_shiny
+is_shiny_runtime <- function(runtime) {
+  !is.null(runtime) && grepl("^shiny", runtime)
+}