+++ R/utils.R
@@ -1,3 +1,6 @@
+# @staticimports pkg:staticimports
+#   is_installed system_file get_package_version %||%
+
 # return a string as a tempfile
 as_tmpfile <- function(str) {
   if (length(str) > 0) {
@@ -24,4 +27,17 @@ knit_devel <- function(input, ...) {
   rmarkdown::render(input,
                     output_options = list(devel = TRUE),
                     quiet = TRUE)
+}
+
+is_accent_color <- function(x) {
+  stopifnot(length(x) == 1)
+  x %in% accent_colors()
+}
+
+accent_colors <- function() {
+  c("primary", "info", "success", "warning", "danger")
+}
+
+dropNulls <- function(x) {
+  x[!vapply(x, is.null, FUN.VALUE = logical(1))]
 }